feat(volar/jsx-directive): support FunctionExpression and ObjectExpression slots
